front: drop importTrainList
