+++ front/public/locales/de/operational-studies.json
@@ -19,8 +19,6 @@
     },
     "failure": "Operation fehlgeschlagen",
     "importTimetable": "Fahrplan importieren",
-    "launchImport": "Import starten",
-    "noResults": "Kein Ergebnis",
     "status": {
       "invalidTimetableItems": "Ungültige Zugfahrpläne.",
       "invalidTimetableItems_one": "Ungültiger Zugfahrplan.",
