+++ front/src/applications/operationalStudies/views/Scenario/components/ImportTimetableItem/ImportTimetableItem.tsx
@@ -1,12 +1,19 @@
-import { useState } from 'react';
+import { useMemo, useState } from 'react';
 
+import { useTranslation } from 'react-i18next';
+
+import { useScenarioContext } from 'applications/operationalStudies/hooks/useScenarioContext';
 import type { TimetableJsonPayload } from 'applications/operationalStudies/types';
+import type { PacedTrain, TrainSchedule } from 'common/api/osrdEditoastApi';
 import { Loader } from 'common/Loaders';
 import { useRollingStockContext } from 'common/RollingStockContext';
+import { useSubCategoryContext } from 'common/SubCategoryContext';
 import type { TimetableItem } from 'reducers/osrdconf/types';
+import { useAppDispatch } from 'store';
 
+import { generateTrainPayloads } from './helpers/generatePayloads';
+import { postFullImportPayload } from './helpers/postPayloads';
 import ImportTimetableItemConfig from './ImportTimetableItemConfig';
-import ImportTimetableItemTrainsList from './ImportTimetableItemTrainsList';
 
 type ImportTimetableItemProps = {
   upsertTimetableItems: (timetableItems: TimetableItem[]) => void;
@@ -21,16 +28,50 @@ const ImportTimetableItem = ({ upsertTimetableItems }: ImportTimetableItemProps)
 
   const { rollingStocks } = useRollingStockContext();
 
-  return rollingStocks ? (
+  const { t } = useTranslation('operational-studies', { keyPrefix: 'importTrains' });
+  const dispatch = useAppDispatch();
+  const { scenario, sandboxId } = useScenarioContext();
+  const {
+    train_schedules: parsedTrainSchedules,
+    paced_trains: parsedPacedTrains,
+    macro_nodes: macroNodes,
+    macro_notes: macroNotes,
+    round_trips: roundTripsFromJsonData,
+  } = trainsJsonData;
+
+  const subCategories = useSubCategoryContext();
+
+  const { pacedTrainsPayload, trainSchedulesPayload } = useMemo<{
+    pacedTrainsPayload: PacedTrain[];
+    trainSchedulesPayload: TrainSchedule[];
+  }>(
+    () => generateTrainPayloads(parsedPacedTrains, parsedTrainSchedules, subCategories),
+    [parsedPacedTrains, parsedTrainSchedules, subCategories]
+  );
+
+  return !isLoading && rollingStocks ? (
     <main className="import-timetable-item" data-testid="import-timetable-item">
       <ImportTimetableItemConfig
         setIsLoading={setIsLoading}
         setTrainsJsonData={setTrainsJsonData}
       />
-      <ImportTimetableItemTrainsList
-        isLoading={isLoading}
-        trainsJsonData={trainsJsonData}
-        upsertTimetableItems={upsertTimetableItems}
+      <button
+        data-testid="launch-import-button"
+        className="btn btn-primary btn-sm ml-auto"
+        type="button"
+        onClick={() =>
+          postFullImportPayload(
+            sandboxId,
+            scenario.id,
+            [...trainSchedulesPayload, ...pacedTrainsPayload],
+            roundTripsFromJsonData,
+            macroNodes,
+            macroNotes,
+            dispatch,
+            t,
+            upsertTimetableItems
+          )
+        }
       />
     </main>
   ) : (
